backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.4
pymupdf==1.23.22
pytesseract==0.3.10
opencv-python==4.9.0.80
pdf2image==1.17.0
pandas==2.2.1
openpyxl==3.1.2
python-docx==1.1.0
numpy==1.26.4
Pillow==10.2.0
pyHanko[crypto]==0.21.0
edge-tts==6.1.9
gTTS==2.5.0
pydub==0.25.1
aiofiles==23.2.1
pywhatkit==5.4
